rgba value sliders with repetitive code

+++ pro-pain-t-app/src/structs/color.rs
@@ -4,11 +4,11 @@ pub struct Color {
     pub r: u8,
     pub g: u8,
     pub b: u8,
-    pub alpha: u8,
+    pub alpha: f32,
 }
 
 impl Color {
-    pub fn new(r: u8, g: u8, b: u8, alpha: u8) -> Self {
+    pub fn new(r: u8, g: u8, b: u8, alpha: f32) -> Self {
         Self { r, g, b, alpha }
     }
 }
